refactor(ir): format methods
