+++ src/ir/passes/printer.rs
@@ -1,5 +1,5 @@
 use crate::ir::{
-    entities::{FunctionKind, ValueKind},
+    entities::FunctionKind,
     module::{DataFlowGraph, Module},
     passes::GlobalPass,
     types::Type,
@@ -23,7 +23,7 @@ where
         Self { buf }
     }
 
-    pub fn print_module(&mut self, module: &Module) -> io::Result<()> {
+    fn print_module(&mut self, module: &Module) -> io::Result<()> {
         writeln!(self.buf, "# orzir module: {} ", module.name())?;
         writeln!(self.buf)?;
         for name in module.identified_type_layout() {
@@ -48,7 +48,7 @@ where
         Ok(())
     }
 
-    pub fn print_function(&mut self, function: Function, module: &Module) -> io::Result<()> {
+    fn print_function(&mut self, function: Function, module: &Module) -> io::Result<()> {
         let data = module.function_data(function).unwrap();
         let function_name = module.value_name(function.into());
 
@@ -101,221 +101,12 @@ where
         writeln!(self.buf, "}}")
     }
 
-    /// Print the value as operand in the instruction
-    pub fn print_operand(&mut self, value: Value, dfg: &DataFlowGraph) -> io::Result<()> {
-        dfg.with_value_data(value, |data| {
-            if data.kind().is_const() {
-                self.print_local_value(value, dfg)
-            } else {
-                write!(self.buf, "{} {}", data.ty(), dfg.value_name(value))
-            }
-        })
-        .unwrap()
+    fn print_global_value(&mut self, value: Value, module: &Module) -> io::Result<()> {
+        write!(self.buf, "{}", value.format_as_global_value(module)?)
     }
 
-    pub fn print_global_value(&mut self, value: Value, module: &Module) -> io::Result<()> {
-        module
-            .with_value_data(value, |data| {
-                match data.kind() {
-                    ValueKind::Zero => write!(self.buf, "zero"),
-                    ValueKind::Undef => write!(self.buf, "undef"),
-                    ValueKind::Bytes(bytes) => {
-                        // hexidecimal format with little endian
-                        write!(self.buf, "0x")?;
-                        for byte in bytes.iter().rev() {
-                            write!(self.buf, "{:02x}", byte)?;
-                        }
-                        Ok(())
-                    }
-                    ValueKind::Array(elems) => {
-                        write!(self.buf, "[")?;
-                        for (i, elem) in elems.iter().enumerate() {
-                            if i != 0 {
-                                write!(self.buf, ", ")?;
-                            }
-                            self.print_global_value(*elem, module)?;
-                        }
-                        write!(self.buf, "]")
-                    }
-                    ValueKind::Struct(fields) => {
-                        write!(self.buf, "{{")?;
-                        for (i, field) in fields.iter().enumerate() {
-                            if i != 0 {
-                                write!(self.buf, ", ")?;
-                            }
-                            self.print_global_value(*field, module)?;
-                        }
-                        write!(self.buf, "}}")
-                    }
-                    ValueKind::GlobalSlot(slot) => {
-                        let ty = module
-                            .with_value_data(slot.init(), |data| data.ty())
-                            .unwrap();
-                        if slot.mutable() {
-                            write!(self.buf, "global {} = {} ", module.value_name(value), ty)?;
-                        } else {
-                            write!(self.buf, "const {} = {} ", module.value_name(value), ty)?;
-                        }
-                        self.print_global_value(slot.init(), module)
-                    }
-                    _ => panic!("unexpected local value kind when printing global value"),
-                }
-            })
-            .unwrap()
-    }
-
-    pub fn print_local_value(&mut self, value: Value, dfg: &DataFlowGraph) -> io::Result<()> {
-        let data = dfg.local_value_data(value).unwrap();
-
-        match data.kind() {
-            ValueKind::Zero => write!(self.buf, "{} zero", data.ty()),
-            ValueKind::Undef => write!(self.buf, "{} undef", data.ty()),
-            ValueKind::Bytes(bytes) => {
-                // hexidecimal format with little endian
-                write!(self.buf, "{} 0x", data.ty())?;
-                if bytes.is_empty() {
-                    write!(self.buf, "00")?;
-                }
-                for byte in bytes.iter().rev() {
-                    write!(self.buf, "{:02x}", byte)?;
-                }
-                Ok(())
-            }
-            ValueKind::Array(elems) => {
-                write!(self.buf, "{} [", data.ty())?;
-                for (i, elem) in elems.iter().enumerate() {
-                    if i != 0 {
-                        write!(self.buf, ", ")?;
-                    }
-                    self.print_local_value(*elem, dfg)?;
-                }
-                write!(self.buf, "]")
-            }
-            ValueKind::Struct(fields) => {
-                write!(self.buf, "{} {{", data.ty())?;
-                for (i, field) in fields.iter().enumerate() {
-                    if i != 0 {
-                        write!(self.buf, ", ")?;
-                    }
-                    self.print_local_value(*field, dfg)?;
-                }
-                write!(self.buf, "}}")
-            }
-            ValueKind::Alloc(alloc) => {
-                write!(self.buf, "{} = alloc {}", dfg.value_name(value), alloc.ty())
-            }
-            ValueKind::Load(load) => {
-                write!(self.buf, "{} = load {}, ", dfg.value_name(value), data.ty())?;
-                self.print_operand(load.ptr(), dfg)
-            }
-            ValueKind::Cast(cast) => {
-                write!(
-                    self.buf,
-                    "{} = {} {}, ",
-                    dfg.value_name(value),
-                    cast.op(),
-                    data.ty()
-                )?;
-                self.print_operand(cast.val(), dfg)
-            }
-            ValueKind::Store(store) => {
-                write!(self.buf, "store ")?;
-                self.print_operand(store.val(), dfg)?;
-                write!(self.buf, ", ")?;
-                self.print_operand(store.ptr(), dfg)
-            }
-            ValueKind::Binary(binary) => {
-                write!(self.buf, "{} = {} ", dfg.value_name(value), binary.op())?;
-                self.print_operand(binary.lhs(), dfg)?;
-                write!(self.buf, ", ")?;
-                self.print_operand(binary.rhs(), dfg)
-            }
-            ValueKind::Unary(unary) => {
-                write!(self.buf, "{} = {} ", dfg.value_name(value), unary.op())?;
-                self.print_operand(unary.val(), dfg)
-            }
-            ValueKind::Jump(jump) => {
-                write!(self.buf, "jump {}(", dfg.block_name(jump.dst()))?;
-                for (i, arg) in jump.args().iter().enumerate() {
-                    if i != 0 {
-                        write!(self.buf, ", ")?;
-                    }
-                    self.print_operand(*arg, dfg)?;
-                }
-                write!(self.buf, ")")
-            }
-            ValueKind::Branch(branch) => {
-                write!(self.buf, "br ")?;
-                self.print_operand(branch.cond(), dfg)?;
-                write!(self.buf, ", {}", dfg.block_name(branch.then_dst()))?;
-                if !branch.then_args().is_empty() {
-                    write!(self.buf, "(")?;
-                    for (i, arg) in branch.then_args().iter().enumerate() {
-                        if i != 0 {
-                            write!(self.buf, ", ")?;
-                        }
-                        self.print_operand(*arg, dfg)?;
-                    }
-                    write!(self.buf, ")")?;
-                }
-                write!(self.buf, ", {}", dfg.block_name(branch.else_dst()))?;
-                if !branch.else_args().is_empty() {
-                    write!(self.buf, "(")?;
-                    for (i, arg) in branch.else_args().iter().enumerate() {
-                        if i != 0 {
-                            write!(self.buf, ", ")?;
-                        }
-                        self.print_operand(*arg, dfg)?;
-                    }
-                    write!(self.buf, ")")?;
-                }
-                Ok(())
-            }
-            ValueKind::Return(ret) => {
-                write!(self.buf, "ret ")?;
-                if let Some(val) = ret.val() {
-                    self.print_operand(val, dfg)?;
-                }
-                Ok(())
-            }
-            ValueKind::Call(call) => {
-                if !data.ty().is_void() {
-                    write!(self.buf, "{} = ", dfg.value_name(value))?;
-                }
-                write!(
-                    self.buf,
-                    "call {} {}(",
-                    data.ty(),
-                    dfg.value_name(call.callee())
-                )?;
-                for (i, arg) in call.args().iter().enumerate() {
-                    if i != 0 {
-                        write!(self.buf, ", ")?;
-                    }
-                    self.print_operand(*arg, dfg)?;
-                }
-                write!(self.buf, ")")
-            }
-            ValueKind::GetElemPtr(gep) => {
-                write!(
-                    self.buf,
-                    "{} = getelemptr {}, ",
-                    dfg.value_name(value),
-                    gep.ty()
-                )?;
-                self.print_operand(gep.ptr(), dfg)?;
-                for idx in gep.indices() {
-                    write!(self.buf, ", ")?;
-                    self.print_operand(*idx, dfg)?;
-                }
-                Ok(())
-            }
-            ValueKind::Function | ValueKind::GlobalSlot(_) | ValueKind::BlockParam => {
-                panic!(
-                    "function, global slot, and block param should not be printed as local value"
-                );
-            }
-        }
+    fn print_local_value(&mut self, value: Value, dfg: &DataFlowGraph) -> io::Result<()> {
+        write!(self.buf, "{}", value.format_as_local_value(dfg)?)
     }
 }
